requirements.txt
flask
python-dotenv
requests
requests-oauthlib
waitress
pytest
